refactor fuse code
