+++ sparql/fusexml.py
@@ -1,0 +1,14 @@
+import xml.etree.ElementTree as ET
+
+class FuseXml:
+    @staticmethod
+    def fuse(to_data, from_data):
+        pass
+
+    @staticmethod
+    def fuse_append(acc, data):
+        pass
+
+    @staticmethod
+    def get_blank_nodes(blank_nodes_data):
+        pass
